+++ NEWS.md
@@ -6,6 +6,7 @@
 - remove frequency cutoff in getsda and use PMD network clusters analysis to determine the cutoff
 - add merge feature for getcluster and such methods could be used to furthor reduce the GlobalStd peaks
 - remove hmdbp data since pmd network analysis could cover this topic
+- remove the dependance of group for quantitative paired peaks
 
 # pmd 0.1.5